added in pull features

+++ gateComs/DSElement.py
@@ -121,36 +121,35 @@ class element:
         python = sys.executable
         os.execl(python, python, *sys.argv)
 
-    def pullMaster(self,sock):
+    def pullBranch(self,sock,branch):
         #let's call the linux commands to pull the repo down
         #we assume you have an ssh key setup
-        branch = "master"
-        logging.debug("pulling latest repo changes")
-        os.system("cd "+currentDirectory+" && git fetch && git reset --hard origin/"+str(branch)+" && git pull origin "+str(branch)+" && exit")
-        #we need to restart this python script to see the changes
-        self.restartProcess(sock)
-
-    def pullDevelop(self,sock):
-        #let's call the linux commands to pull the repo down
-        #we assume you have an ssh key setup
-        branch = "develop"
         logging.debug("pulling latest repo changes")
         currentDirectory = sys.path[0]
-        os.system("cd "+currentDirectory+" && git fetch && git reset --hard origin/"+str(branch)+" && git pull origin "+str(branch)+" && exit")
+        if(devMode == False):
+            os.system("cd "+currentDirectory+" && git fetch && git reset --hard origin/"+str(branch)+" && git pull origin "+str(branch)+" && exit")
+        else:
+            print("cd "+currentDirectory+" && git fetch && git reset --hard origin/"+str(branch)+" && git pull origin "+str(branch)+" && exit")
         #we need to restart this python script to see the changes
         self.restartProcess(sock)
 
     def shutdown(self,sock):
         #let's call the linux commands to shutdown the pis
         logging.debug("shutting down Pis...")
-        sock.close()
-        os.system("sudo shutdown now")
+        if(devMode == False):
+            sock.close()
+            os.system("sudo shutdown now")
+        else:
+            print("sudo shutdown now")
 
     def reboot(self,sock):
         #let's call the linux commands to shutdown the pis
         logging.debug("rebooting Pis...")
-        sock.close()
-        os.system("sudo reboot now")
+        if(devMode == False):
+            sock.close()
+            os.system("sudo reboot now")
+        else:
+            print("sudo reboot now")
 
     def runProgram(self,sock,LED):
         gate = DSUtils.Gate(sock,(self.serverAddress,self.port),"rainbow")
@@ -178,27 +177,29 @@ class element:
                 body = data['body'] #the body of the message
                 recipient = data['recipient'] #the intended recipient. If there isn't one, the message is for everyone
                 logging.debug("recieved data")
-                self.currentColor = body
                 if(subject == "disconnect"):
                     logging.debug("we recieved a disconnect request")
                     break;
                 if(subject == "updateColor"):
+                    self.currentColor = body
                     logging.debug("updating color: "+str(self.currentColor))
                     if(devMode == False):
                         LED.customColor(body)
-                    logging.debug("done updating")
                 if(subject == "updateAnimation"):
+                    self.currentColor = body
                     logging.debug("updating animation: "+str(self.currentColor))
                 if(subject == "systemCommand"):
-                    logging.debug("updating color: "+str(self.currentColor))
-                    if(devMode==False):
-                        if(self.body=="shutdown"):
-                            self.shutdown()
-                        if(self.body=="reboot"):
-                            self.reboot()
-                    else:
-                        if(self.currentColor=="update"):
-                            self.pullDevelop(sock)
+                    command = body['command']
+                    arguments = body['arguments']
+                    logging.debug("performing command: "+command)
+                    if(command=="shutdown"):
+                        self.shutdown()
+                    if(command=="reboot"):
+                        self.reboot()
+                    if(command=="update"):
+                        branch = arguments[0]
+                        self.pullBranch(sock,branch)
+
         logging.debug("disconnected")
